+++ Assets/Scripts/FollowCamera.cs
@@ -4,15 +4,27 @@ using UnityEngine;
 
 public class FollowCamera : MonoBehaviour
 {
-    // Start is called before the first frame update
+    public Transform target;
+    float offsetX;
+    float offsetY;
+
     void Start()
     {
-        
+        if (target == null)
+            return;
+
+        offsetX = transform.position.x - target.position.x;
+        offsetY = transform.position.y - target.position.y;
     }
 
-    // Update is called once per frame
     void Update()
     {
-        
+        if (target == null)
+            return;
+
+        Vector3 pos = transform.position;
+        pos.x = target.position.x + offsetX;
+        pos.y = target.position.y + offsetY;
+        transform.position = pos;
     }
 }
